A simple script for animating slide out menus, info panes, and contact forms.
This script executes a simple sliding animation on elements with the
class "center-tray" and the IDs "left-tray" and "right-tray". By
default, the script will slide the first div with class "center-tray"
and position absolute 512px left or right. If you put a fixed or
absolute positioned div underneath your center tray you can give your
users quick and easy access to contact forms, navigation menus, or
information pages anywhere on your website like you see in the below
(WIP) example:

http://p4p.dantruittht.ml/

By changing the animationlength variable and the properties of the trays
object you can easily extend the functionality of this script to a
handful of other applications.

Animationlength will fairly obviously change how long the animation
lasts.

trays.moveleft determines whether the left tray (#left-tray) will move
or not. This property is set to "off" by default, but can be set to "on"
create navigation menus that slide over the pages main content, or
panels that move with the rest of the site content.

trays.left will set the width of the left tray. When moveleft is "off"
trays.left will determine how far other elements will slide. If the
specified value is larger than the device's viewport, this property will
automatically be set to the device's width.

trays.movecenter is set to "on" by default. This property determines
whether the div with class "center-tray" will move. Set this property to
"off" to prevent the center tray from moving when you want your menus
and forms to slide over the site's main content.

trays.moveright is analogous to trays.moveleft, but on the right side.
This property affects the div with ID right-tray.

trays.right is analogous to trays.left.

Animations are triggered when a user clicks on an element with one of
the following IDs:

open-left-tray
close-left-tray
open-right-tray
close-right-tray

The function of these IDs should be fairly evident.

I put this script together because I found that I tend to use these
kinds of effects a lot and I figured it would be beneficial for me to
have one easily modified script that I could use to implement these
features without wasting time writing code that is functionally
identical to other code I've written. As interest in this script builds
or as I use this particular script more often you can expect it to
become increasingly flexible and easier to customize.